Convert JdbcTemplate to NamedParameterJdbcTemplate for prevent SQL Injection

+++ SpringBootRestAPI/src/main/java/com/ranacorporation/SpringBootRestAPI/business/repository/StudentRepository.java
@@ -9,7 +9,7 @@ public interface StudentRepository {
 
     Student findById(long id);
 
-    int save(Student student);
+    long save(Student student);
 
     int update(long id, Student student);
